requirements.txt
openai>=0.28.0
google-generativeai
pillow
pymupdf
tenacity
python-dotenv 